+++ Civi/DataProcessor/FilterHandler/MultipleFieldFilter.php
@@ -182,7 +182,7 @@ class MultipleFieldFilter extends AbstractFilterHandler {
       }
       $dataFlow = $dataSource->ensureField($fieldSpec);
       if ($dataFlow && $dataFlow instanceof SqlDataFlow) {
-        $clauses[] = new SqlDataFlow\SimpleWhereClause($dataFlow->getName(), $fieldSpec->name, $filterParams['op'], $filterParams['value'], $this->fieldSpecification->type);
+        $clauses[] = new SqlDataFlow\SimpleWhereClause($dataFlow->getName(), $fieldSpec->getName(), $filterParams['op'], $filterParams['value'], $this->fieldSpecification->type);
       }
     }
     if (count($clauses)) {